Fix b3_utils library to work properly in workspace

+++ backend/lib/b3_operations/src/pending/state.rs
@@ -20,7 +20,7 @@ impl OperationState {
         id
     }
 
-    pub fn request_counter(&self) -> usize {
+    pub fn request_counter(&self) -> u64 {
         self.nonce.current()
     }
